Removed old commented code
Removed some old bit of code which were commented out. One of these was a long-ago commented out section in the programme form view, which lists the custom fields. This has however had the side-effect of breaking submission of the form (which is presumably why that bit of code had been commented out!)