Add Simple_Admin_Controller
Simple_Admin_Controller is an abstraction to avoid code-repetition of a
number of ways of doing simple CRUD style administration interfaces we
do generally.

It abstracts the functions. All that needs done is to extend the class
and define some simple variables.

What is more, the introduction of Simple_Admin_Controller extends the
test coverage since a single test can cover a number of controllers.