Override delete() in Image to soft-delete the actual image files.
When an image was deleted only the database record was set to hidden - the actual files were
left where they were.

This meant that if the URLs were known (e.g. google cache / image search) then the images could still
be displayed.

This fix prefixes the image and thumbnail filenames with a '.' so that the old URLs no longer work.

We could actually delete the image files, but as "deleting" only hides the database records it makes
more sense to hide the files so we can retrieve them if need be.